Move claim anchor actions to its own subfolder
