+++ src/main/kotlin/dev/mizarc/bellclaims/application/actions/claim/anchor/BreakClaimAnchor.kt
@@ -1,11 +1,11 @@
-package dev.mizarc.bellclaims.application.actions.claim
+package dev.mizarc.bellclaims.application.actions.claim.anchor
 
 import dev.mizarc.bellclaims.application.persistence.ClaimFlagRepository
 import dev.mizarc.bellclaims.application.persistence.ClaimPermissionRepository
 import dev.mizarc.bellclaims.application.persistence.ClaimRepository
 import dev.mizarc.bellclaims.application.persistence.PartitionRepository
 import dev.mizarc.bellclaims.application.persistence.PlayerAccessRepository
-import dev.mizarc.bellclaims.application.results.claim.BreakClaimAnchorResult
+import dev.mizarc.bellclaims.application.results.claim.anchor.BreakClaimAnchorResult
 import dev.mizarc.bellclaims.domain.values.Position3D
 import java.util.UUID
